+++ topobank/manager/signals.py
@@ -63,8 +63,8 @@ def remove_files(sender, instance, **kwargs):
             _log.info(f'Deleting {datafile.name}...')
             datafile.delete()
         except Exception as exc:
-            _log.warning("Topography id %d, attribute '%s': Cannot delete data file '%s', reason: %s",
-                         instance.id, datafile_attr_name, datafile.name, str(exc))
+            _log.warning(f"Topography id {instance.id}, attribute '{datafile_attr_name}': Cannot delete data file "
+                         f"{datafile.name}', reason: {str(exc)}")
 
     datafile_path = instance.datafile.name
     squeezed_datafile_path = instance.squeezed_datafile.name
@@ -80,15 +80,15 @@ def remove_files(sender, instance, **kwargs):
     # Otherwise we abort deletion.
     if datafile_path is not None and not datafile_path.startswith(instance.storage_prefix):
         _log.warning(f'Datafile is stored at location {datafile_path}, but storage prefix is '
-                     f'{instance.storage_prefix}. I will not attempt to deleting everything at this prefix.')
+                     f'{instance.storage_prefix}. I will not attempt to delete everything at this prefix.')
         return
     if squeezed_datafile_path is not None and not squeezed_datafile_path.startswith(instance.storage_prefix):
         _log.warning(f'Squeezed datafile is stored at location {squeezed_datafile_path}, but storage prefix is '
-                     f'{instance.storage_prefix}. I will not attempt to deleting everything at this prefix.')
+                     f'{instance.storage_prefix}. I will not attempt to delete everything at this prefix.')
         return
     if thumbnail_path is not None and not thumbnail_path.startswith(instance.storage_prefix):
         _log.warning(f'Thumbnail is stored at location {thumbnail_path}, but storage prefix is '
-                     f'{instance.storage_prefix}. I will not attempt to deleting everything at this prefix.')
+                     f'{instance.storage_prefix}. I will not attempt to delete everything at this prefix.')
         return
     recursive_delete(instance.storage_prefix)